Potions instead of chests
The existence of chests implies the existence of items, and to be honest I don't wanna create that many items. I'm already thinking of what the PCs and enemies could wield, but I just don't wanna make inventory items, so unless there's a strong opinion, the potions can be for stat changes, and maybe "break" or "drink" them when you move over them on a tile.