Test Scene Updates
Made an Immovables interface to host items such as the potions and other scenery, as well as a Potion script to mark what is a potion and what will be scenery. Those items should be able to be walked through unless given the impassable script.
I'm about to change the sizes of the potions because they are MASSIVE, so I'll update those.
I can also implement a stat change script, so I'm waiting on the units to get a Stats implementation (I don't know how to do it with scriptable objects, and I need to make enemies first)