Added more inventory functionality
Set the first two slots to default to Physical And Magical attacks, as well as a Use() method. The idea is that the UI would remove the potions but not the attacks, but pass all of the strings to the Use() method. Still waiting on learning how to get a targeted enemy